requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
backoff==2.2.1
beautifulsoup4==4.12.0
blinker==1.5
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
cryptography==40.0.2
ctranslate2==3.10.3
dataclasses-json==0.5.7
decorator==4.4.2
deepspeech==0.9.3
dnspython==2.3.0
duckdb==0.7.1
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.95.1
faster-whisper==0.3.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.10.7
firebase-admin==6.1.0
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.13.3
hyperdb-python==0.1.2
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.135
llvmlite==0.39.1
loguru==0.7.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@6dea21fd7f7253bfe450f1e2512a0fe47ee2d258
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pinecone-client==2.2.1
posthog==3.0.1
proglog==0.1.10
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
pydub==0.25.1
pyFLAC==2.1.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==12.1.3
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.3
selenium-stealth==1.0.6
semantic-version==2.10.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
setuptools-rust==1.5.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
SQLAlchemy==1.4.47
starlette==0.26.1
streamlit==1.20.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.28.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.2
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.2.3
whisper==1.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
zstandard==0.21.0
